Split utilities method into MathUtils and OIUtils
Cherry-picked from commit 9ffb9c79
