+++ src/main/java/org/team1540/rooster/drive/pipeline/SimpleJoystickInput.java
@@ -1,8 +1,9 @@
 package org.team1540.rooster.drive.pipeline;
 
 import edu.wpi.first.wpilibj.Joystick;
-import org.team1540.rooster.Utilities;
 import org.team1540.rooster.functional.Input;
+import org.team1540.rooster.util.MathUtils;
+import org.team1540.rooster.util.OIUtils;
 
 /**
  * Simple tank-style input from a WPILib {@link Joystick}. The left and right joysticks are used to
@@ -26,23 +27,23 @@ public class SimpleJoystickInput implements Input<TankDriveData> {
 
   @Override
   public TankDriveData get() {
-    double triggerValue;
-    if (fwdAxis != -1 && backAxis != -1) {
-      triggerValue = Utilities.processDeadzone(joystick.getRawAxis(fwdAxis), deadzone)
-          - Utilities.processDeadzone(joystick.getRawAxis(backAxis), deadzone);
-    } else {
-      triggerValue = 0;
-    }
-    double leftThrottle = Utilities.constrain(
-        Utilities.processDeadzone(
-            Utilities.invertIf(invertLeft, joystick.getRawAxis(leftAxis)), deadzone
-        ) + triggerValue, 1);
-    double rightThrottle = Utilities.constrain(
-        Utilities.processDeadzone(
-            Utilities.invertIf(invertRight, joystick.getRawAxis(rightAxis)), deadzone
-        ) + triggerValue, 1);
+      double triggerValue;
+      if (fwdAxis != -1 && backAxis != -1) {
+          triggerValue = OIUtils.processDeadzone(joystick.getRawAxis(fwdAxis), deadzone)
+              - OIUtils.processDeadzone(joystick.getRawAxis(backAxis), deadzone);
+      } else {
+          triggerValue = 0;
+      }
+      double leftThrottle = MathUtils.constrain(
+          OIUtils.processDeadzone(
+              MathUtils.negateDoubleIf(invertLeft, joystick.getRawAxis(leftAxis)), deadzone
+          ) + triggerValue, 1);
+      double rightThrottle = MathUtils.constrain(
+          OIUtils.processDeadzone(
+              MathUtils.negateDoubleIf(invertRight, joystick.getRawAxis(rightAxis)), deadzone
+          ) + triggerValue, 1);
 
-    return new TankDriveData().withAdditionalFeedForwards(leftThrottle, rightThrottle);
+      return new TankDriveData().withAdditionalFeedForwards(leftThrottle, rightThrottle);
   }
 
   /**
@@ -98,8 +99,7 @@ public class SimpleJoystickInput implements Input<TankDriveData> {
    * axis.
    * @param invertLeft Whether to invert the axis value of the left axis.
    * @param invertRight Whether to invert the axis value of the right axis.
-   * @param deadzone The deadzone for the axes (see {@link Utilities#processDeadzone(double,
-   * double)}).
+   * @param deadzone The deadzone for the axes (see {@link OIUtils#processDeadzone(double, double)}).
    */
   public SimpleJoystickInput(Joystick joystick, int leftAxis, int rightAxis, int fwdAxis,
       int backAxis, boolean invertLeft, boolean invertRight, double deadzone) {
