+++ src/main/java/org/team1540/rooster/util/robots/PIDTuningRobot.java
@@ -9,9 +9,9 @@ import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.team1540.rooster.Utilities;
 import org.team1540.rooster.preferencemanager.Preference;
 import org.team1540.rooster.preferencemanager.PreferenceManager;
+import org.team1540.rooster.util.OIUtils;
 import org.team1540.rooster.wrappers.ChickenTalon;
 
 /**
@@ -151,8 +151,8 @@ public class PIDTuningRobot extends IterativeRobot {
       if (enablePID) {
         motor1.set(controlModeChooser.getSelected(), setpoint);
       } else {
-        motor1
-            .set(ControlMode.PercentOutput, Utilities.processDeadzone(joystick.getRawAxis(1), 0.1));
+          motor1
+              .set(ControlMode.PercentOutput, OIUtils.processDeadzone(joystick.getRawAxis(1), 0.1));
       }
     }
   }
